Create readme
Confirmed Cases in tableau is used for better visualization
The time series data of confirmed COVID-19 cases is used for studying how people on Twitter  would behave with the increasing number of confirmed cases. We obtain the data from Kaggle website which takes data collectively from Indian Ministry of Health & Family Welfare, covid19india.org . 
The date range we used is from Jan 30th to June 16th. The dataset used for this is attached in "covid_19_india" file.